test: make sure APIClass return expected type results
