+++ apps/meteor/app/api/server/index.ts
@@ -52,4 +52,4 @@ import './v1/moderation';
 // This has to come last so all endpoints are registered before generating the OpenAPI documentation
 import './default/openApi';
 
-export { API, APIClass, defaultRateLimiterOptions } from './api';
+export { API, defaultRateLimiterOptions } from './api';
